cleanup some js and fix comment collapse button
- extracts some duplicate js code into common modules

- breaks out the latex auto-render code into its own module

- fixes the collapse comment button to hide the comment vote controls
  as well
