+++ app/javascript/packs/pageload.js
@@ -1,0 +1,10 @@
+export function onPageLoad(callback) {
+  if (
+      document.readyState === "complete" ||
+      (document.readyState !== "loading" && !document.documentElement.doScroll)
+  ) {
+    callback();
+  } else {
+    document.addEventListener("DOMContentLoaded", callback);
+  }
+}
