+++ app/javascript/packs/submission_actions.js
@@ -1,3 +1,6 @@
+const pageload = require("./pageload");
+const throttle = require("lodash/throttle");
+
 function updateHideOpacity(element, json) {
   if (json.status === "hidden") {
     element.style.opacity = "0.5";
@@ -31,22 +34,31 @@ function handleResponse(element, a, response) {
   }
 }
 
+function fetchActionResponse(element, a, kind) {
+  fetch("/users/submission_actions", {
+    method: "PUT",
+    headers: {
+      "Content-Type": "application/json",
+      "X-CSRF-Token": document.getElementsByName("csrf-token")[0].content
+    },
+    body: JSON.stringify({
+      submission_action: {
+        kind: kind,
+        submission_short_id: element.id
+      }
+    })
+  }).then(response => handleResponse(element, a, response));
+}
+
 function setOnClick(element, a, kind) {
+  let throttledRequest = throttle(
+    fetchActionResponse,
+    750,
+    { 'trailing': false }
+  );
   a.addEventListener("click", function(e) {
     e.preventDefault();
-    fetch("/users/submission_actions", {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json",
-        "X-CSRF-Token": document.getElementsByName("csrf-token")[0].content
-      },
-      body: JSON.stringify({
-        submission_action: {
-          kind: kind,
-          submission_short_id: element.id
-        }
-      })
-    }).then(response => handleResponse(element, a, response));
+    throttledRequest(element, a, kind);
   });
 }
 
@@ -55,19 +67,10 @@ function setupActionLinks(element) {
   setOnClick(element, element.getElementsByClassName("save")[0], "saved");
 }
 
-const callback = function() {
+pageload.onPageLoad(function() {
   const elements = document.getElementsByClassName("submission");
 
   for (let i = 0; i < elements.length; i++) {
     setupActionLinks(elements[i]);
   }
-};
-
-if (
-    document.readyState === "complete" ||
-    (document.readyState !== "loading" && !document.documentElement.doScroll)
-) {
-  callback();
-} else {
-  document.addEventListener("DOMContentLoaded", callback);
-}
+});
